ai/llm-rag/with-nvidia/requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-nvidia-ai-endpoints
langchain-text-splitters
faiss-cpu
pypdf
openai
